Rework how if statements are lowered.
Previously we tried to preserve the if-elseif-else structure all the way from source code to IR to lowered code. This added complexity, and made it difficult to ensure else-if conditions are only evaluated if the initial condition was false.

Now IR only supports if-else, which is both simpler to generate in the frontend and simpler to lower in the backends.